[Feature #162380953]Finish user sign-up
