+++ .gitignore
@@ -4,3 +4,4 @@
 .vscode/
 venv/
 .pytest_cache/
+.cache
